Move get_public_pages_queryset outside of api.v2
